Move AdminExample base class to client/admin subfolder since that's the only place that uses it.
