+++ examples/client/admin/__init__.py
@@ -1,0 +1,7 @@
+from ... import Example
+
+
+class AdminExample(Example):
+    def __init__(self):
+        # TODO: admin user doesn't have enough permissions
+        super().__init__(user="admin", password="admin")
