+++ examples/client/admin/query_roles.py
@@ -15,7 +15,7 @@
 # limitations under the License.
 ################################################################################
 
-from ... import AdminExample
+from . import AdminExample
 
 
 class QueryRoles(AdminExample):
